// src/types/user.ts
export interface User {
  id: number;
  name: string;
  email: string;
  type: string;
  password: string;
  about: string;
  address: string;
  city: string;
  country: string;
  createdAt: string;
  experience: string;
  firstname: string;
  lastname: string;
  mobile: string;
  qualification: string;
  updatedAt: string;
}